Add change in existing VM image type
